change: move components into core folder
