+++ src/core/Billboard.tsx
