+++ src/core/CurveModifier.tsx
