+++ src/core/Detailed.tsx
