+++ src/core/FlyControls.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import { ReactThreeFiber, useThree, useFrame, Overwrite } from 'react-three-fiber'
 import { FlyControls as FlyControlsImpl } from 'three/examples/jsm/controls/FlyControls'
-import useEffectfulState from './helpers/useEffectfulState'
+import useEffectfulState from '../helpers/useEffectfulState'
 
 export type FlyControls = Overwrite<
   ReactThreeFiber.Object3DNode<FlyControlsImpl, typeof FlyControlsImpl>,
