+++ src/core/Stars.tsx
