+++ src/core/Text.tsx
