+++ src/core/softShadows.tsx
