+++ src/core/useCamera.tsx
