+++ src/core/useCubeTexture.tsx
