+++ src/core/useDetectGPU.tsx
