+++ src/core/useGLTF.tsx
