+++ src/core/useNormalTexture.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import { useTexture } from './useTexture'
 import { RepeatWrapping, Texture, Vector2 } from 'three'
 
-import { normalsList } from './helpers/normal-assets'
+import { normalsList } from '../helpers/normal-assets'
 
 const NORMAL_ROOT = 'https://rawcdn.githack.com/emmelleppi/normal-maps/f24c810fc1d86b5b1e5dfea914b668f70b5f2923'
 const DEFAULT_NORMAL = normalsList[0]
