--- src/index.tsx
@@ -1,68 +1,0 @@
-// Abstractions
-export * from './Billboard'
-export * from './Detailed'
-export * from './Line'
-export * from './PositionalAudio'
-export * from './Text'
-export * from './Environment'
-export * from './Effects'
-
-// Cameras
-export * from './OrthographicCamera'
-export * from './PerspectiveCamera'
-
-// Controls
-export * from './DeviceOrientationControls'
-export * from './FlyControls'
-export * from './MapControls'
-export * from './OrbitControls'
-export * from './TrackballControls'
-export * from './TransformControls'
-export * from './PointerLockControls'
-
-// Loaders
-export * from './Loader'
-export * from './useCubeTexture'
-export * from './useFBX'
-export * from './useGLTF'
-export * from './useProgress'
-export * from './useTexture'
-
-// Misc
-export * from './Html'
-export * from './meshBounds'
-export * from './Reflector'
-export * from './Shadow'
-export * from './Stats'
-export * from './useAspect'
-export * from './useCamera'
-export * from './useDetectGPU'
-export * from './useHelper'
-export * from './useContextBridge'
-export * from './useAnimations'
-export * from './useFBO'
-
-// Modifiers
-export * from './useSimplification'
-export * from './useSubdivision'
-export * from './useTessellation'
-export * from './useEdgeSplit'
-export * from './CurveModifier'
-
-// Shaders
-export * from './MeshDistortMaterial'
-export * from './MeshWobbleMaterial'
-export * from './shaderMaterial'
-export * from './Sky'
-export * from './softShadows'
-export * from './Stars'
-export * from './ContactShadows'
-
-// Shapes
-export * from './shapes'
-export * from './RoundedBox'
-export * from './ScreenQuad'
-
-// Prototyping
-export * from './useMatcapTexture'
-export * from './useNormalTexture'
